improve package structure
