+++ application/src/main/java/org/opentripplanner/framework/snapshot/persistence/repository/timetable/MutableTimetableRepo.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.framework.snapshot.persistence.timetable;
+package org.opentripplanner.framework.snapshot.persistence.repository.timetable;
 
 import org.opentripplanner.framework.snapshot.domain.timetable.TimetableRepo;
