+++ application/src/main/java/org/opentripplanner/framework/snapshot/persistence/repository/timetable/TimetableConfig.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.framework.snapshot.persistence.timetable;
+package org.opentripplanner.framework.snapshot.persistence.repository.timetable;
 
 public class TimetableConfig {
