+++ application/src/main/java/org/opentripplanner/framework/snapshot/persistence/snapshot/DefaultTransactionalContext.java
@@ -4,7 +4,7 @@ import org.opentripplanner.framework.snapshot.application.TransactionalContext;
 import org.opentripplanner.framework.snapshot.domain.timetable.TimetableRepo;
 import org.opentripplanner.framework.snapshot.domain.transfer.TransferRepo;
 import org.opentripplanner.framework.snapshot.domain.DomainEvent;
-import org.opentripplanner.framework.snapshot.persistence.world.MutableTransitWorld;
+import org.opentripplanner.framework.snapshot.persistence.snapshot.world.MutableTransitWorld;
 
 public class DefaultTransactionalContext implements TransactionalContext {
